[PR] 1.1.0 rust

+++ src/directory_flattener.py
@@ -45,9 +45,11 @@ def directory_flattener(input_path: str, output_path: str, file_extension: str) 
                 for file in filenames:
                     if file.endswith(file_extension):
 
+                        # Get abspath from root
                         # Prepare relative paths:
-                        relative_dir = os.path.relpath(root, maybe_dir)
-                        relative_file = os.path.join(relative_dir, file)
+                        root_abspath = os.path.join(os.path.abspath(root), file)
+                        input_abspath = os.path.abspath(input_path)
+                        relative_file = root_abspath.removeprefix(input_abspath)
 
                         # Get unique filename:
                         unique_filename = uuid.uuid4().hex
@@ -57,9 +59,14 @@ def directory_flattener(input_path: str, output_path: str, file_extension: str) 
                         )
 
                         current_file = os.path.abspath(os.path.join(root, file))
+                        current_file = "\\\\?\\" + current_file
 
                         # Copying files:
-                        shutil.copy(current_file, new_path_and_filename)
+                        if os.path.exists(current_file):
+                            shutil.copy(current_file, new_path_and_filename)
+                        else:
+                            print(len(current_file))
+                            print("err")
 
                         # Add to a mapping
                         dir_structure_mapping[unique_filename_with_ext] = relative_file
